+++ includes/class-github-api.php
@@ -39,7 +39,7 @@ class GitHub_Updater_GitHub_API {
 	 *
 	 * @var integer
 	 */
-	 protected static $hours;
+	 public static $hours = 1;
 	 
 
 	/**
@@ -51,7 +51,7 @@ class GitHub_Updater_GitHub_API {
 	 */
 	public function __construct( $type ) {
 		$this->type = $type;
-		self::$hours = apply_filters( 'github_updater_set_transient_hours', 1 );
+		self::$hours = apply_filters( 'github_updater_set_transient_hours', self::$hours );
 	}
 
 	/**
@@ -121,23 +121,24 @@ class GitHub_Updater_GitHub_API {
 	 * @since 1.0.0
 	 */
 	public function get_remote_info( $file ) {
-		$remote = get_site_transient( md5( $this->type->repo . $file ) );
+
+		$remote = get_site_transient( 'ghu-' . md5( $this->type->repo . $file ) );
 		if ( ! $remote ) {
 			$remote = $this->api( '/repos/:owner/:repo/contents/' . $file );
 
 			if ( $remote ) {
-				set_site_transient( md5( $this->type->repo . $file ), $remote, ( self::$hours * HOUR_IN_SECONDS ) );
+				set_site_transient( 'ghu-' . md5( $this->type->repo . $file ), $remote, ( self::$hours * HOUR_IN_SECONDS ) );
 			}
 		}
 
 		$this->type->branch = $this->get_default_branch( $remote );
 
 		if ( ! $remote ) return;
+		$this->type->transient = $remote;
 		preg_match( '/^[ \t\/*#@]*Version\:\s*(.*)$/im', base64_decode( $remote->content ), $matches );
 
 		if ( ! empty( $matches[1] ) )
-			$this->type->remote_version = $matches[1];
-
+			$this->type->remote_version = trim( $matches[1] );
 	}
 
 	/**
@@ -175,13 +176,13 @@ class GitHub_Updater_GitHub_API {
 	 * @return string latest tag.
 	 */
 	public function get_remote_tag() {
-		$response = get_site_transient( md5( $this->type->repo . 'tags' ) );
+		$response = get_site_transient( 'ghu-' . md5( $this->type->repo . 'tags' ) );
 
 		if ( ! $response ) {
 			$response = $this->api( '/repos/:owner/:repo/tags' );
 
 			if ( $response ) {
-				set_site_transient( md5( $this->type->repo . 'tags' ), $response, ( self::$hours * HOUR_IN_SECONDS ) );
+				set_site_transient( 'ghu-' . md5( $this->type->repo . 'tags' ), $response, ( self::$hours * HOUR_IN_SECONDS ) );
 			}
 		}
 
@@ -202,7 +203,6 @@ class GitHub_Updater_GitHub_API {
 		$newest_tag     = $tags[ $newest_tag_key ];
 
 		$this->type->newest_tag    = $newest_tag;
-//		$this->type->download_link = $this->type->uri . '/archive/' . $this->type->newest_tag . '.zip';
 		$this->type->tags          = $tags;
 	}
 
@@ -232,15 +232,17 @@ class GitHub_Updater_GitHub_API {
 	 * @since 1.9.0
 	 * @return base64 decoded CHANGES.md or false
 	 */
-	public function get_remote_changes() {
+	public function get_remote_changes( $changes ) {
+		if ( ! class_exists( 'Markdown_Parser' ) )
+			require_once 'markdown.php';
 
-		$remote = get_site_transient( md5( $this->type->repo . 'changes' ) );
+		$remote = get_site_transient( 'ghu-' . md5( $this->type->repo . 'changes' ) );
 
 		if ( ! $remote ) {
-			$remote = $this->api( '/repos/:owner/:repo/contents/CHANGES.md' );
+			$remote = $this->api( '/repos/:owner/:repo/contents/' . $changes  );
 
 			if ( $remote ) {
-				set_site_transient( md5( $this->type->repo . 'changes' ), $remote, ( self::$hours * HOUR_IN_SECONDS ) );				
+				set_site_transient( 'ghu-' . md5( $this->type->repo . 'changes' ), $remote, ( self::$hours * HOUR_IN_SECONDS ) );				
 			}
 		}
 		
@@ -252,7 +254,64 @@ class GitHub_Updater_GitHub_API {
 			}
 			$this->type->sections['changelog'] = $changelog;
 		}
+	}
+	
+	/**
+	 * Read the repository meta from GitHub API
+	 *
+	 * Uses a transient to limit calls to the API
+	 * @since 2.2.0
+	 * @return base64 decoded repository meta data
+	 */
+	public function get_repo_meta() {
+		$remote = get_site_transient( 'ghu-' . md5( $this->type->repo . 'meta' ) );
+		$meta_query = '?q=' . $this->type->repo . '+user:' . $this->type->owner;
 
+		if ( ! $remote ) {
+			$remote = $this->api( '/search/repositories' . $meta_query );
+
+			if ( $remote ) {
+				set_site_transient( 'ghu-' . md5( $this->type->repo . 'meta' ), $remote, ( self::$hours * HOUR_IN_SECONDS ) );				
+			}
+		}
+
+		if ( ! $remote ) return false;
+		
+		$this->type->repo_meta = $remote->items[0];
+		$this->add_meta_repo_object();
+	}
+
+	/**
+	 * Add remote data to type object
+	 *
+	 * @since 2.2.0
+	 */
+	private function add_meta_repo_object() {
+		$this->type->last_updated = $this->type->repo_meta->pushed_at;
+		$this->type->rating = $this->make_rating();
+		$this->type->num_ratings = $this->type->repo_meta->watchers;
+	}
+
+	/**
+	 * Create some sort of rating from 0 to 100 for use in star ratings
+	 * I'm really just making this up, more based upon popularity
+	 *
+	 * @since 2.2.0
+	 * @return integer
+	 */
+	private function make_rating() {
+		$watchers    = $this->type->repo_meta->watchers;
+		$forks       = $this->type->repo_meta->forks;
+		$open_issues = $this->type->repo_meta->open_issues;
+		$score       = $this->type->repo_meta->score; //what is this anyway?
+
+		$rating = round( $watchers + ( $forks * 1.5 ) - $open_issues );
+
+		if ( 100 < $rating ) {
+			return 100;
+		}
+
+		return $rating;
 	}
 
 }